[FIX] account_sepa_direct_debit: fix default mandate identifier for windows
When creating a mandate, it needs to have a unique identifier.
To do so, it sets the default identifier to be essentially the timestamp down to
the microsecond.
However the date formatting was set using %s (Unix timestamp) instead of %S
(seconds), which had the side-effect of crashing on non-Unix compliant OS's
(such as Windows).

opw 1858850