[FIX] account_3way_match: use service product in tests
When splitting purchase from stock, 'account_3way_match' dependencies
looked good, as the business code was not using stock models. However,
this was not the case for the test cases ! This commit makes the test
use service product (instead of consummable) which is natively supported
in purchase module (as we don't want to create a bridge module between
stock and account_3way_match).
Before this commit, starting test of account_3way_match without stock
installed obviously crashes.