[IMP] pos_loyalty: spend and earn points with identical precision
Set-up the following.
- Loyalty Program:
    Points per currency: 1.00
    Points Rounding:     0.01
    Applies on Product A
- Reward:
    Reward Type: Discount
    Discount:    100 %
    Point Cost:  1.00

Then, in the POS:
- Buy Product A costing 10.50 => you receive 10.50 points
- Buy Product B costing 20.00, apply the reward

Only 10 € are substracted, not 10.50 €.

It is more logical that you can spend the reward with the same precision
that it is earned. It is always possible to define a 'Points Rounding'
to 1.00 if the user wants to take advantage of the extra cents on which
the points won't be applied.

opw-1861455