[IMP] l10n_mx_edi_external_trade: Support the dest address in the CFDI
When a sale is an external trade, we could have this case:

1.- I generate an invoice to my customer in Belgium,
2.- That products are sent to the USA, not to Belgium.

In this case I need indicate that my customer is in Belgium but the
destination address is in the USA.

When the sale module is installed, the shipping address is used,
otherwise the delivery address of the contact is used. When the shipping
contact is the same than the invoice contact, no destination address is
specified.

opw-1859102
closes #2364