[FIX] l10n_mx_edi: Do not allow remove a certificate used
Improved that not allow remove a Certificate that was used to sign an
invoice.

Why?

When an invoice is signed is saved the Certificate, because is used if
the invoice is cancelled, if the record is removed when try cancel We
get the next error:

```
ValueError: Expected singleton: l10n_mx_edi.certificate()
```

closes odoo/enterprise#2354