[FIX] web_studio: remove drop leftovers
When drag & dropping a field on the view (list or form) very sides, the drop
handling was not correctly executed and the dropped element was stucked.

This is due to the fact that the drop handling was set on the `drop` event,
which was not triggered when dropping the columns on the very sides.

The drop handling for moving fields is now bound on the `stop` event of the
draggable element, which is triggered even though the element is dropped outside
of the view.