[FIX] sale_subscription: do not put a notebook in a group
In rev. odoo/enterprise@a47cf36 the notebook of subscription form view has been
put inside a <group> element.

The reason behind that was to remove the border-bottom in the form view (which
is set because the notebook is not the last child).

Putting the notebook inside a group causes hidden issues (like right padding,
enable to click on the tab in Studio, etc.).

Even if this is probably not the best way to remove this border, we put the
notebook inside a div instead of a group to avoid side effects.