[FIX] web_studio: move a field inside a created page
Before this ref, after creating a new page and moving a field inside it (more
precisely, inside a group of this page), the field was wrongly moved after
the group, not inside it.

The xpath generated by a moved field can only be placed as a direct child of
another xpath. The issue here is that the current generated xpath was considered
as valid because its parent node was inside the current xpath.

A condition has been added to create a new xpath if the moved field is not a
direct child.

Co-authored-by: David Monjoie <dmo@odoo.com>