[IMP] web_grid: widget in grid view
This commit add the supports of 'light' widget in the grid
view. For now, widget were not supported in Grid view, only
formatting (like 'float_time') is handled.
This commit provides a first step to support widget, with the following
principes :
1/ widget is responsible of formatting the value, and displaying the
content in grid cell. The widget manage the format to allow using options
propagate in the code directly (rather than on the field node of the
view definition).
2/ widget implements its own behavior (handlers, ...), and triggers up
'grid_cell_edited' event (with the formatted value) when the value need
to be committed (adjustment).
3/ if a widget is indicated on the `field` tag in view definition, and no
grid widget is implemented, the formatting method will be deduced from
the traditionnal field widget registry.

This commit also introduces as the `float_toggle` as example of this
new API.

Task #39079