[IMP] website_sale_dashboard: improve KPIs use of website_sale
Website dashboard is modified in website_sale_dashboared module. In order
to make the same update in digest so we override the action url computation
of digest. This view is displayed when you click on 'view more' link of
eCommerce section in digest emails.

See commit https://github.com/odoo/odoo/commit/8eec939a7885c860c6c821298709e567bd625027 in community for more information.

This commit is linked to task ID 30655 and PR #2180.