[IMP] currency_rate_live: currency order in banxico
Banxico has made changes in its WebService and now the expected order of
returned currencies has changed.

FYI it was:

 USD SAT > USD FIX > EUR > CAD > JPY > GBP

and now is:

 EUR > CAD > USD FIX > GBP > JPY > USD SAT

The first order was expected so all currency rates wre mixed up.

For this reason, the code is change to use the `IDSERIE` identifying the
currency and not simply the order they might appear by.

Document with the identifiers corresponding to currencies:
http://www.banxico.org.mx/ley-de-transparencia/consultas-frecuentes/{960A6514-B048-02B8-4BF2-920034786674}.pdf

With this change, we also use USD FIX instead of USD SAT.

closes #2366