[FIX] helpdesk: access ticket with token display messages
Have a ticket issued by a portal user.
Access this ticket **while unlogged** with the ticket token controller
/helpdesk/ticket/id/token

Before this commit, the discussion on the ticket's chatter weren't displayed
This was because the access token was not passed on to JS, then to the message controller

After this commit, the discussion is displayed.

OPW 1858890
closes #2370