[FIX]currency_rate_live:  Multicompany awareness when updating currency
In case of multicompany instance _update_currency_banxico method only creates a rate for
the admin's company and the other currency rates get deleted in the update, resulting on
the other company rates not being udpated. In any other time the admin changes company
and then when the cron for updating the rates runs the company rates for that company only gets updated.

This change avoids the encapsulated method update_rate from removing other companies currency rates.

opw:1866488