[IMP] web: comparisons in dashboard view
A previous commit [odoo/333ac27cd5c23e3e1a4f7d78828f28a2c844f8f4] has introduced a new widget
TimeRangeMenu that allows to create two sets of data to
compare relatively to a given date or datetime field.

This commit allows the dashboard view to render
appropriately two sets of data. In comparison mode, the
subviews are instantiated in comparison mode and the
statistics are rendered differently. The variation
expressed in percentage is presented above the values
comming from the data and comparison data (in
that order) separated by a 'vs'.

Task ID: 1835644