[FIX] web_studio: add an image/priority on kanban editor
After the DataManager refactoring (see odoo/odoo@cb33984), the field name was
not accessible inside the field description anymore.

We used to access it from the field description at several places (mostly in
sidebar and kanban record editor).

The case was already managed for ViewEditorManager (and thus for sidebar) with
rev odoo/enterprise@271378a in `_processFields`, but not for kanban record
editor.

However for kanban record editor, we only need field name and label (string)
which are used in a dialog for adding Priority / Image field.

This commit fixes the traceback by correctly passing minimal required values to
the dialog.

Closes #2380
Task : 1863059

Co-authored-by: nla-odoo <nla@odoo.com>