[IMP] web_unsplash: fit Odoo js coding & keep cache (hack)
This commit adapt JS code to fit Odoo guidelines.

It also introduce a `hack` to prevent the UnsplashAPI to be destroyed every
time the media dialog is closed.
Indeed, UnsplashAPI has a cache system to recude unsplash call, it is then
better to keep its state to benefic from it from one media dialog call to
another.

Unsplash API will either be (it's still being discussed):
* a service (ideally coming with an improvement to not auto load the service)
* initialized in the website_root (trigger_up)

Thanks to QSM

task-35060