[IMP] web_unsplash: make api key readable not only by admin
As the key was retrieve with a RPC reading ir.config_parameter, limited user
could not read the key (eg: website editor). It would throw a read error when
opening the media dialog as only sudo can read this.

Now, we check that the user has the rights to access unsplash key and retrieve
it as sudo afterward.

The rights check is implemented as a method on res.users to be easily
overridable by other modules to give access to unsplash key.
It is the case for website designer.

At some point, `project` module will also give unsplash key access to project
manager to be able to select a project image cover. (task-1848170)

  - web_unsplash (OE): base.group_erp_manager got key access rights
  - website (OC): website.group_website_designer got key access rights
  - project (OC): not implemented yet, some users will also get the rights

Note: Since unsplash module is from enterprise, the solution to create an
unsplash group would not work.
Indeed, community modules like website would need to set that unsplash group
when you have website editor group (implied_ids).
As unsplash is enterprise, the group would not be found in community.

task-35060