[FIX] l10n_mx_edi: Fix cases with multi-currency on payment.
When the payment complement is generated in MXN to an invoice in USD the
exchange rate set in the detail must be the rate in the payment date and
not in the invoice date, by the next reason:

1. I have the next invoice
  Amount total = 100 USD
  1 USD = 20 MXN
  Total MXN = 2000

2. When I make the payment:
  1 USD = 18 MXN
  Then, the Amount Paid is 1800 MXN that is the same that 20 USD

To do not have any stamp error the `TipoCambioDR` that is the rate in
the invoice document must be 18.

Now the CFDI have the next value in this case:

```<Pago10:Pago FechaPago="2018-09-05T12:00:00" FormaDePagoP="01"
MonedaP="MXN" Monto="1800.00" NumOperacion="INV/2018/0008">
<Pago10:DoctoRelacionado MetodoDePagoDR="PPD"
IdDocumento="670FB007-56C8-49DD-9343-AF56F7CA5717" Serie="INV/2018/"
Folio="8" MonedaDR="USD" TipoCambioDR="18.000000" NumParcialidad="1"
ImpSaldoAnt="100.00" ImpPagado="100.00" ImpSaldoInsoluto="0.00"/>
</Pago10:Pago>```

Note: The PAC verify that `ImpPagado` * `TipoCambioDR` <= `Monto`

And now work correctly.

Also was improved that work when the company is configured to work with
3 decimals or more.