[FIX] sign: Don't reduce pdf size on responsible modification
PDF size should not be reduced on share wizard andwhen change
a responsible for a field, it should be updated

Currently, when change a responsible for a specific field, its not updating and when
open 'send' or 'share' wizard then PDF size reducing because while opening a
wizard, a PDFframe reinitialization process again start. So here we've a
new PDFframe in 'iframeWidget' and that's why newly added fields not added in the template.

To fix these both issue, improved condition just by added one more condition
i.e if 'modal-dialog' class found then not create a new instance of 'EditablePDFIframe'.

Related to Task: 1862371