[FIX] stock_barcode: update O-CMD.SAVE in demo pdf
In 10.0 with a5031e4b erroneous O-BTN.save was replaced with O-CMD.SAVE
and some PDF with barcodes were updated.

But there was still  stock_barcode/static/description/barcodes_demo_data.pdf
that was not updated and now is with this commit.

For general knowledge, this had to be done with:

  SBDIR=/home/odoo/src/enterprise/10.0/stock_barcode
  sudo sh -c 'echo a4 > /etc/papersize'
  sh <(cat $SBDIR/doc/make_barcodes.sh | sed 's/-p A4//')
  ps2pdf <(cat barcodes_actions.ps $SBDIR/doc/barcodes_demo_data.ps) $SBDIR/static/description/barcodes_demo_data.pdf
  rm barcodes_actions.ps barcodes_demo_data.ps

with:

- barcode 0.99 (https://www.gnu.org/software/barcode/)
- ghostscript 9.23 (https://www.ghostscript.com/)
- libpaper 1.1.24 (https://packages.debian.org/unstable/source/libpaper)

installed.

note: not needed for 11.0 since the issue was solved in 4f23241d9

opw-1860964
closes #2426