[IMP] sign: Display directly the reponsible dropdown on signature fields
Purpose
=======

Displaye directly a dropdown while selecting a responsible

When selecting a responsible on a signature item, don't force the user to type at least 1 character but display directly a dropdown.

Also while typing, a dropdown position not maintaining that's why removed that portion which is adding extra top position.

Related to Task: 1869556