[IMP] web_studio: report editor
The report editor is a major update on the actual report editor.
Before this commit, the report editor allowed to edit the static text of the existing reports.

After this commit, the report editor allows the creation, deletion and update of the reports.
Reports can be updated by:
* changing the structure of the report by adding and removing content and nodes at any level
* editing the content
    - changing the text of a node
    - drag and dropping new content to a report (static or dynamic) from a list of pre-defined building block, using realistic preview
    - adding or changing which field is bound to any the content of the report, using a field selector
    - changing the layout of every element of the report (limited to width, margins, background and font color, and css class)
    - adding multi column structure to the content
    - when adding dynamic content, changing the widget used for the rendering as well as the options of the widgets
* adding content level security
    - allow to assign visibility group to any element of the report (and their children)
    - allow to assign visibility conditions to any element, using a domain selector

Technical notes:
---------------
The rendering of the report preview happens server side, and is loaded into an iframe inside the editor workspace.
When rendering the preview, the server adds meta information on every node containing
* the id of the template the node is originating from in ´data-oe-id´
* the xpath from the origin of the template that generated the node in ´data-oe-xpath´
* the context available at the moment of the evaluation of the node, so we know on what model and variables it can be bound in ´data-oe-context´

The report preview is loaded in an iframe so that it has its own style, without interference of the style of the web client or studio itself.
This way, its layout is closer to the actual server rendered report.

The reports are composed of many templates: the actual template of the reports used t-call directives to append itself to the layout,
and other t-call to include content or other templates. When an edition is done, we extend with a inherited view for each template that has changes.

We have refactored studio in 2 ways:
* split the source in multiple folders: common, reports, views
* introduce the AbstractEditorManager that has as responsibility to orchestrate the creation of the editor space and sidebar,
to refresh them when needed and to handle high level events. It has to concrete inheritance: the ViewEditorManager for the everything but the report editor,
and the ReportEditorManager for the report editor.

Co-authored-by: Martin Geubelle <mge@odoo.com>
Co-authored-by: Christophe Matthieu <chm@odoo.com>
Co-authored-by: Vincent Schippefilt <vsc@odoo.com>