[FIX] account_reports: communication followup multiline
On the follow-up report, a very long line communication (for example
from a very long "Reference/Description" on an invoice) can break the
width of the table.

It might be alright in the interface, since this should be rare, there
is a horizontal scrollbar and the interface is less ugly. But on the PDF
there is no usable scrollbar so part of the information may be out of
page and lost.

note: a similar fix is in 10.0 with #2449

opw-1861726
closes #2454