[FIX] project_timesheet_synchro: avoid security error on res.user
This fixes an issue in multi-company mode when 'share partners to all
companies' option is unchecked.
In this case, the record rule named 'res.partner company' is active and
prevents users from a company to access partners data from the other
companies.

This explains why a security error is raised  when a user whose related
partner is from company 'A' tries to synchronize Awesome Timesheet and
a task from company 'A' is assigned to a partner from company 'B'.
It's not allowed to read the name of this partner.

This fix allows tasks to be synchronized even if the partner assigned
isn't accessible.

This is the easier fix as 'user_id' is never used.

opw-1866793