[FIX] mrp_account: Wrong MO cost analysis
When the module mrp_byproduct is installed, it is possible to add
other finished products in a BOM. These other finished products can
be expressed in an other uom's category than the BOM.

So it was wrong to sum all the finished in the cost analysis report.

PS: there is no module bridge between mrp_byproduct and mrp_account.

opw:1868436