[FIX] web_cohort: allow many2one as measure in cohort view
A previous commit [e1aebdccc405c6e26bc189294198cac4e10d96e8] allows the incorporation of a cohort
view in a dashboard view. This commit fixes a problem
introduced then. If an aggregate with associated field is of
type many2one, the dashboard view instantiate its subviews
with that field added as measure in the measures list.
So the cohort view should be capable to render itself in case
a measure of type many2one is selected. This is now possible.