[FIX] mrp_account: cost analysis pdf ok
The cost analysis report on product.template and product.product was
corrected in 2856173 so it works on product.template model.

But this still happen:

- using button "Cost Analysis" on product template or product.product
  then "Print" will either have the erroneous filename, or an error
  saying the record doesn't exist or has been deleted.

- on product.template using the print menu and "Cost Structure Analysis"
  report will try to show the cost analysis of the product.product with
  ID being the product.template ID.

The first issue happen because when printing the PDF, the file name is
based on [object].name, but object is:
  product.template({ID of a product.product})

mismatching IDs of product.template and product.product models.

The second issue happened because the print menu use the current ID or
selected IDs, the the report expecting a product.product->id receive a
product.template->id.

By changing the model of the report to `product.product`, both issues
are fixed, and it is more logical to have the report being on
product.product since it is what the report is printing.

opw-1863039
closes #2462