[IMP] iot: add camera functionality on workorder
When you take a picture and a device is connected,
the picture will be taken from the camera on the iot box instead
 of the system asking you for a local file.

The data will be retrieved in base64 from the iotbox.
For that we needed to override the widget of the picture field.
When a an iot device is connected, it should not send the html
form with the local file on it, but immediately save
the picture data in the value of the field itself.