[FIX] website_sign: fix select2 configuration not properly reset
The configuration of the select2 change handler is reset on setting
a value for the field, at which point it is regenerated to contain
the new updated values if the user wants to select another one.

This configuration is a variable set on a function, so it is global
and is not destroyed alongside any widget when exiting the edit mode.

So if, at that point, you exit the edit mode, by going back to the
"Document Templates" view for example then re-enter the edit mode
of this document, the configuration has not been reset and still
contains a reference to the previous edit session parties.

Thus, the first select will trigger the process_party function of
the previous configuration and update the "parties" object of the
previous edit session rather than the new one. When the code later
tries to read from the "parties" object, it fails since it has not
been properly updated since the update was triggered on an old
object rather than the new one.

To fix this, we need to get rid of the old configuration. Either
we do it when we exit edit mode (I didn't find how to do this) or
we make sure to reset the configuration when we enter edit mode,
which is what this fix is doing.