[FIX] stock_barcode: hide information button while loading
In the barcode client action, confirming an inner form view will destroy 
it and reload all the lines. The issue is: before all the lines are 
rendered, we can click on the information button (i). This button will 
load the information form view. Then, once all the lines are rendered 
the client action will overide the information view. At this time, the 
user cannot do anything good unless quitting the barcode app. 

This commit hide the information button while the client action load and 
render all the lines

Task : 1857705