[FIX] project_timesheet_forecast: Fix traceback on report view
Purpose
=======

When opening the 'Timesheets & Forecast Analysis' menuitem, there is
a traceback due to the bad definition of the groupby filters.

Specification
=============

Instead of having special groupbys A > B and B > C, define 3 filters A, B, C
to allow any combination the user really wants.

This commit is related to bug#1871683 .