[IMP] mail_enterprise: pager in side attachment preview
Before this commit, it was not possible to display several
attachments in previewer. It was just displaying the first
image/pdf from multiple attachments.

After this commit, if several images or pdfs are attached to a
document, the attachment previewer will display next/previous
arrows to switch between these attachments.

This is related to task ID 1844685
Closes #2221