[NEW] account_invoice_extract: new iap module to extract invoices fro…
…m pdf and png

This module send your invoices in pdf/png to iap to extract data from the files.
The extracted data will be used to prefill an invoice (in_invoice or out_refund only)

To send your files, create your invoice and add the file in a message in the chatter.
Your attachment must be the first in the chatter to be sent to iap

Depending on your configuration for your company, the files are automatically, manually
or not sent. By default, files are sent manually.

You can too send your files by email using the existing automatic invoice system
to create an invoice and send attachment to iap.

This module change the pdf/png preview on your invoices to show you the different
extracted fields on your attachment. You can directly click on this fields
to set the value in your invoice (if the invoice is still in draft)
Set the value by clicking on the right box in the pdf will improve better the ocr
service than write directly the right value in the invoice.