[FIX] account_reports: leap day and last year filter
The current calculation when calculating a financial report over the
previous year is:

- if we are in a leap year, substract 366 days
- if we are not in a leap year, substract 365 days

This doesn't take into account if we are after or before 29th february.

Thus, we get this result getting the previous year from these dates:

- January 1st 2017 -> January 2nd 2016 (we substracted 365 days)
- January 1st 2016 -> December 31th 2014 (we substracted 366 days)

When we would have expected:

- January 1st 2017 -> January 1st 2016 (we substracted 366 days)
- January 1st 2016 -> January 1st 2015 (we substracted 365 days)

This commit uses dateutile.relativedelta which takes the leap day into
account. There is an argument that 29th february minus one year should
be 1st march if it is a date_from, but we just get 28th in any case
(since having a fiscal year on 29th february would not be wise).

opw-1872275
closes #2516