[FIX] web_studio: tests: skip test due to broken grouped layout
The intent of a recent commit is to improve layout of grouped lists [1].
However, the layout is a bit broken [2].
Some tests broke as a result [3].

There are issues to revert the 1st commit (lots of conflicts).
Also, we think that it goes in the correct direction with making grouped
list layout nicer, so it seems better to fix/improve on top of that.

Task-ID 1874605

[1] https://github.com/odoo/odoo/commit/f7cce79cfe3509955048aca2424bbc17b0867577
[2] https://drive.google.com/file/d/1X0cC09EggUSxatT3sgN1UpeYgXLsWL3f/view?usp=sharing
[3] https://github.com/odoo/enterprise/commit/3ec7d3e16f8a363470092301b6a54ef0870726c4