[FIX] account_reports: keep line return in outlook
In at least up to outlook 2007, using the CSS:

  white-space: pre-wrap;

doesn't seem to be supported.

Thus if the followup report was received on outlook, we would possibly
get something very messy without any newline.

This commit uses:

  white-space: pre; white-space: pre-wrap;

so we fallback on `white-space: pre` on outlook. Outlook also seems to
wrap lines even if it should not which is what we want (and why we used
pre-wrap instead of pre).

note: this is for 10.0 up to not-including 11.0 (whose fix is in #2519)

opw-1871122
closes #2518