[REM] website_version: remove module
This module can be partially replaced with the new model website_page
introduced in v11 and the 'publish/unpublish' feature.
You can easily create new page side by side and improve it without
publish them and link the page to the new view to make it the master.

This module conflict with multi website and is too fragile.
It can easily break anothers inherit.

AB Testing (e.g. show 1 time on 3 a special page) will be probably reimplemented
in the near futur with a more robust solution.