[FIX] web_studio: raise error on groups without xmlid
To add a group to the visibility of a given field, the group needs to have an
external xml_id.
This is the case if the group is native or has been created by studio.
It isn't the case for groups created by hand, and can thus create a traceback
while trying to use these groups in studio.

opw 1871722