[FIX] sign: do not create a sign request forcing a sms_number
The sms_number is a related, not mandatory field on a sign.request

In hr_contract_salary for instance, the sms_number is not given (KeyError).

In the wizard it was given but just the value of related.
This was making an unecessary write on the partner when the key was given.