[FIX] web_studio: set a group on a report node
There were multiple issues while trying to set a group on a node in the
report editor.

Firstly the operation was not correct, a wrong override of `_triggerViewChange`
in the component was creating a badly formatted operation.

Secondly the user context (in which the `studio` key is added) was not correctly
passed when fetching the report views. We have taken the opportunity to add this
context in multiple places where it was missing.

Finally the `groups` attribute in arch was not processed when fetching report
views. One could set this attribute on a node like `<span groups='base.group_system'>`.
In `load_views` this attribute is popped so an override of `_apply_group` in
`ir.ui.view` is done to process the node and set the attribute `studio_groups`
on the node. Something similar has to be done for report views (which do not use
the same view mechanism but rely on qweb rendering).