[REF] web_studio: simplify ViewEditorManager tests
When `mockRPC` was set in a ViewEditorManager test, the rpc `get_default_value`
has to be overriden, which is quite annoying.

This has been simplified by putting a default answer in the mock server.

The helper function to create a ViewEditorManager in tests has also been moved
in test_utils file and a bit refactored.