[FIX] stock_barcode: avoid unecessary read()
Scanning a destination location in the stock_barcode app means the flow 
is ended. The client action doesn't know immediately the current barcode 
represent a location. It make an rpc first to check if the barcode is a 
package one. We would like to avoid making useless rpcs.  

This commit invert the order. First we check if the barcode is a 
destination location. Then if it's a package.