[FIX] project_timesheet_synchro:tests: no running services in tests suites
Related to following community PR: https://github.com/odoo/odoo/pull/26519

We do not want to deploy services in tests environments, because it
may influence the execution of the tests. For instance, if we receive
a new direct message in Odoo, we do not want to have a chat window
appear in those environments.

Timesheet UI is a service provider. So for any test on timesheet UI,
we should disable its logic as a Service Provider.

Closes #2553