[FIX] account_reports: newline summary outlook 365
The summary field in a financial report is text but newlines are shown
thanks to CSS (white-space: pre-wrap) that can not work on some HTML
clients.

In #2518 this was solved for Outlook 2007 and over, but it was still
broken with Office 365 which seems to totally strip the white-space CSS.

With this commit, we use the text widget that escape the content and add
HTML `<br>` to newlines.

note: this is for 10.0 up to not-including 11.0 (whose fix is in #2554)

opw-1871122
closes #2552