[FIX] account_reports: newline summary outlook 365
The summary field in a financial report is text but newlines are shown
thanks to CSS (white-space: pre-wrap) that can not work on some HTML
clients.

In #2519 this was solved for Outlook 2007 and over, but it was still
broken with Office 365 which seems to totally strip the white-space CSS.

With this commit, we use the text widget that escape the content and add
HTML `<br>` to newlines.

There was also an issue with the editing part: when we saved a summary
it was saved as text, but displayed as html (so saving `<hello>`
would be saved and sent as `&lt;hello&gt;` but was displayed as HTML
`<hello>` until the page was refreshed).

note: 11.0 equivalent of #2552

opw-1871122
closes #2554