[FIX] account_reports: fix cash flow statement for financing activities
Previously, the cash flow statement was relying only on the income and expense accounts (account.account.type) to determine positive cash flow or negative cash flow. This was only partially correct as some cash flows (like issuance of payment of debts) are not involving any income/expense account.

To have correct figures in all use cases, a complete rework of the report was needed. We now have a dedicated pre-query aliasing the account_move_line table in case we're in the cash flow statement report.

Note also that the module account_reports should be upgraded, as modifications were made in the report lines data, to benefit from this patch.

Was task: 31220
Was PR #2502