[FIX] account_3way_match: Compare field 'date_due' with 'date' object, instead of datetime
An example, to understand:
Current date is 2018-08-23 ; current time is 15:46:00. I have an invoice with due_date= 2018-08-23. datetime.today()'s result contains the current time (not just the date), so before this fix, we had 2018-08-23 < datetime.today(), and the invoice would be  considered as 'to pay' even though it will only be overdue tomorrow.