[FIX] web_studio: wait for the field to be renamed in tour
There was a nondeterministic error in multiple runbot builds in the Studio tour
"test_new_app". It was due to a step being executed too early in the tour.

A field label is changed (which triggers a field renaming). Then the next step
(add a selection field) requires to click on the 'Add' sidebar tab. However,
after a node modification, the node is selected back to update the 'Options'
sidebar tab with the new modification (default behavior in Studio).

Depending on the server response time for the field renaming and view edition,
the result selected tab was either 'Add' or 'Options'. To avoid this
nondeterministic state, an `extra_trigger` has been added in the tour to wait
until the operation is applied before switching on 'Add' tab.