[FIX] project_forecast: avoid retriggering compute on saving settings
Saving the res.config.settings will call the `create` method and
propagate the new value on the res.company for the related
fields, like 'forecast_uom'.
Even the value has not changed, writing the value will be done and if
some computed fields (like resource_time on project.forecast)
depends on this company field, a recomputation of all the record
will be done (which might be very slow depending of the number
of database entries).

This commit prevent writing on 'forecast_uom' field on the company by
not giving to the create of res.config.settings.

Note: the same change was already done in accounting at
c781c1e5e2d186b1c03c5b60e522f8e730d59a0e