[FIX] web_studio: edit/create a report without record
It's not possible to render a report without any linked record.
For this reason, a warning is displayed when the user tries to edit a report
if there is no record on the related model.

However, there was a traceback in this case, which was not intended, obviously.

The `on_attach_callback` has been protected in case the editor is not displayed.

Morever, in the mentioned case, a white screen was displayed (the
ReportEditorAction was empty). Now, a `history_back` is performed to display
back the available reports.