[FIX] account_asset: domain in multicompany
In a multicompany environment, an user who has access to multiple companies is
able to select accounts and analytic accounts of multiple companies. This can
easily lead to mistakes in the definition of the asset category.

With this commit the user can only read/select accounts and analytic accounts
belonging to the company already set in the asset category.

This is what is already done in most invoice or other accounting views

Closes odoo/odoo#26578

Cherry-pick of odoo/odoo@5de5953ae7d60748ae7d7a2a66437d63fa8bec3f