[FIX] web_studio: undo the correct failed operation
In case of slow network, the automatic undo of a fail operation is not correct
and leads to wrong undone operations.

Let's imagine the following use case:
 (1) the user changes something
 (2) the user changes something else before (1) is done
 (3) the user changes something else before (1) is done

with (1) which will trigger a server error.

In this case, the operations are wrong.

Remember that the operations stack JS side is synchronous but the EditorManagers
use a MutexedDropPrevious. Let's resume what different states after each
operation should be:

Op      Editor stack  MDP       Network
----------------------------------------
(1)do   [1]           blocked   RPC_1 [1]
(2)do   [1, 2]        pending
(3)do   [1, 2, 3]     pending

                                *RPC_1 fails*

(4)undo [2, 3]        released  RPC_2 [2, 3]

Before this rev. the Editor state after (4) was [1, 2] instead of [2, 3] because
when performing a undo, the last operation was pop instead of the failed one.

To solve this issue, an internal `id` has been set on each operation, which is
used to undo the correct failed operation.

Note that this only works if the operation fails server-side. If the operation
fails at rendering time, the previous behaviour (the last operation on the
stack) still applies.