[FIX] web_studio: display 'Add' tab in report editor
When editing a report, the 'Add' tab should be the one selected (just like
when editing Views) to encourage the user to add stuff.

However, when editing the ir.actions.report record in the 'Report' tab, the
whole Editor is reloaded (because the paperformat could have changed for example).

In this case, to avoid flickering, the 'Report' tab is still selected.