[FIX] web_cohort: Fixed traceback while clicking Period / Count cell
Cohort view now supports comparison mode with help of TimeRangeMenu widget.
However, after introducing this feature, a traceback is faced when we click
on Period or Count Column of any data row in a regular cohort view without
comparison. This is because of improper data-type (attribute which helps
deciding type of the cohort view) being set in xml file.

This commit fixes the traceback by properly setting data-type attribute
on the data row, and also updates related test cases.

Task 1880288
Closes #2584